Hunter notes, preloading data
Preloading data on initialisation for other readers
Added hunter note pool and preview handling
Added out of range protection for grouping and ordering arrays
All sorts of refactoring and moving code around without actually changing how it works
